requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
statsmodels
python-dotenv
openai
requests
joblib
pydantic
streamlit